% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abin.R
\name{abin}
\alias{abin}
\title{Prints out a barplot & returns a barplot of the given result.}
\usage{
abin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of iterations which are sampled.}

\item{n}{The size of the table.}

\item{p}{The Probability of the 1, event.}
}
\value{
Prints out a barplot & returns a barplot of the given result.
}
\description{
Prints out a barplot & returns a barplot of the given result.
}
\examples{
abin(iter = 100, n = 10, p = 0.5)
}
