% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cltu.R
\name{cltu}
\alias{cltu}
\title{Generates a sample and displays it to the user given various inputs to generate good data.}
\usage{
cltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{size of sample}

\item{iter}{number of iterations}

\item{a}{the size of the lower limit}

\item{b}{the size of the upper limit}
}
\value{
A histogram plot to show the distribution of sample generated.

A vector of values from the distribution
}
\description{
Generates a sample and displays it to the user given various inputs to generate good data.
}
\examples{
cltu(n=20,iter=100000)
}
