% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBoot2.R
\name{myBoot2}
\alias{myBoot2}
\title{Generates a sample and displays it to the user given various inputs to generate good data.}
\usage{
myBoot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{the vector of values}

\item{fun}{function type (mean, sd, median, etc.)}

\item{alpha}{alpha value}

\item{cx}{a value which changes the cex value in the text()}

\item{...}{of values which can be entered}
}
\value{
A histogram plot to show the values along with the interval.

an interval, function type, and vector.
}
\description{
Generates a sample and displays it to the user given various inputs to generate good data.
}
\examples{
myBoot2(x=c(1,2,3,4,2,2,3,3,2,3,2,1))
}
