% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlNorm.R
\name{mymlNorm}
\alias{mymlNorm}
\title{Finds the optimal amount of plane tickets to sell on a flight given that some passengers won't show up.}
\usage{
mymlNorm(x, mu, sig, ...)
}
\arguments{
\item{x}{The values of the distribution}

\item{mu}{The mu value of the normal distribution.}

\item{sig}{The sigma value of the normal distribution.}

\item{...}{More values which can be entered}
}
\value{
Makes a plot of all of the values in terms of a normal distribution

Returns a list of values such as the max likely-hood, coordinates, and values.
}
\description{
Finds the optimal amount of plane tickets to sell on a flight given that some passengers won't show up.
}
\examples{
mymlNorm(x=c(5,7,7,8,10),mu=seq(5,10,length=1000),sig=seq(0.1,4,length=1000),lwd=2,labcex=1)
}
