% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Make a graph curve of pnorm.}
\usage{
myncurve(a, mu, sigma)
}
\arguments{
\item{a}{The input for where the pnorm ends from negative infinity to a.}

\item{mu}{The mu value which controls the mean.}

\item{sigma}{The sigma value which controls standard deviation.}
}
\value{
a curve of of the given pnorm in the form of Y(X<=a).
}
\description{
Make a graph curve of pnorm.
}
\examples{
myncurve(12, 10, 2)
}
