% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zR.R
\name{myz}
\alias{myz}
\title{Make a vector of z values}
\usage{
myz(x)
}
\arguments{
\item{x}{A vector of numeric values.}
}
\value{
A vector of z values.
}
\description{
Make a vector of z values
}
\examples{
myz(1:4)
}
